+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### **0.8.0** (2022-01-xx)
-* xx
+* Address book - Your Addresses: allow to filter change addresses
+* Address book - Your Addresses: switch between addresses balance or label
 
 ### **0.7.9** (2022-01-11)
 * Allow keys to be exported in address book